git-svn-id: https://ccmc-software.googlecode.com/svn@1278 acdf4c94-f73e-48c1-5845-2db45fb2849b

+++ kameleon-plus/src/ccmc/Adapt3D.h
@@ -25,8 +25,8 @@ namespace ccmc
 {
 	struct SmartGridSearchValues
 	{
-		double xl_sg,xr_sg,yl_sg,yr_sg,zl_sg,zr_sg;
-		double dx_sg,dy_sg,dz_sg;
+		float xl_sg,xr_sg,yl_sg,yr_sg,zl_sg,zr_sg;
+		float dx_sg,dy_sg,dz_sg;
 		int *indx;
 		int *esup1;
 		int *esup2;
